Auth

+++ src/main/java/hyve/petshow/controller/representation/ServicoRepresentation.java
@@ -1,14 +1,10 @@
 package hyve.petshow.controller.representation;
 
-
-import hyve.petshow.domain.Servico;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 public class ServicoRepresentation {
-    private Long id;
+    private Integer id;
 	private String nome;
     private String descricao;
 }
